adding the skills info
