+++ src/Components/RecentJobs/index.jsx
@@ -1,6 +1,7 @@
 import RecentBox from "./RecentBox"
 
 function RecentJobs() {
+    
 
     return (
         <div class='bg-white'>
